refactor: update skapi import paths and remove deprecated admin module
