+++ src/code/admin.ts
@@ -1,6 +1,3 @@
-import { Skapi } from 'skapi-js';
-
-export const skapi = new Skapi(import.meta.env.VITE_ADMIN, 'skapi', { autoLogin: window.localStorage.getItem('remember') === 'true' }, JSON.parse(import.meta.env.VITE_ETC));
 
 export let dateFormat = (timestamp: number) => {
     let currentDate = new Date(timestamp);
