+++ src/components/navBar.vue
@@ -76,7 +76,7 @@ nav#navBar(ref="navBar")
 <script setup lang="ts">
 import { useRoute, useRouter } from "vue-router";
 import { onMounted, onBeforeUnmount, ref } from "vue";
-import { skapi } from "@/code/admin";
+import { skapi } from "@/main";
 import { loginState, user, updateUser, customer } from "@/code/user";
 import { showDropDown } from "@/assets/js/event.js";
 import { setAutoHide, removeListener } from "./navBar-autohide.ts";
