+++ src/views/account-setting.vue
@@ -75,7 +75,7 @@ Modal(:open="proceedVerification" @close="proceedVerification=false")
 </template>
 
 <script setup lang="ts">
-import { skapi } from '@/code/admin';
+import { skapi } from '@/main';
 import { user, updateUser, emailSubscribed } from '@/code/user';
 import router from '@/router';
 import { computed, ref, nextTick, watch } from 'vue';
