+++ src/views/change-password.vue
@@ -148,7 +148,7 @@ br
 <script setup lang="ts">
 import { useRoute, useRouter } from 'vue-router';
 import { ref, nextTick } from 'vue';
-import { skapi } from '@/code/admin';
+import { skapi } from '@/main';
 import { user } from '@/code/user';
 const router = useRouter();
 const route = useRoute();
