+++ src/views/delete-account.vue
@@ -48,7 +48,7 @@ br
 </template>
 
 <script setup lang="ts">
-import { skapi } from '@/code/admin';
+import { skapi } from '@/main';
 import { useRoute, useRouter } from 'vue-router';
 import Checkbox from '@/components/checkbox.vue';
 import { serviceList } from './service-list';
