+++ src/views/service/mail.vue
@@ -259,11 +259,11 @@ import type { ComputedRef, Ref } from "vue";
 import { currentService, serviceAutoMails } from "./main";
 import { user } from "@/code/user";
 import Code from "@/components/code.vue";
-import { dateFormat } from "@/code/admin";
+import { dateFormat } from "@/main";
 import Table from "@/components/table.vue";
 import Modal from "@/components/modal.vue";
 import Pager from "@/code/pager";
-import { skapi } from "@/code/admin";
+import { skapi } from "@/main";
 import { devLog } from "@/code/logger";
 import Select from "@/components/select.vue";
 import Toggle from "@/components/toggle.vue";
